Separate graph.helper into helper.js and renderer.jsx
