+++ .gitignore
@@ -70,4 +70,5 @@ coverage
 sandbox/rd3g.sandbox.bundle.js.map
 gen-docs
 .DS_Store
-.vscode/
+**/.DS_Store
+.vscode
